refactoring of resolution parameters
