+++ include/exadg/compressible_navier_stokes/solver.h
@@ -26,8 +26,9 @@
 #include <exadg/compressible_navier_stokes/driver.h>
 
 // utilities
+#include <exadg/operators/resolution_parameters.h>
+#include <exadg/time_integration/resolution_parameters.h>
 #include <exadg/utilities/general_parameters.h>
-#include <exadg/utilities/resolution_parameters.h>
 
 // application
 #include <exadg/compressible_navier_stokes/user_interface/declare_get_application.h>
@@ -42,7 +43,7 @@ create_input_file(std::string const & input_file)
   GeneralParameters general;
   general.add_parameters(prm);
 
-  SpatialResolutionParameters spatial;
+  SpatialResolutionParametersMinMax spatial;
   spatial.add_parameters(prm);
 
   TemporalResolutionParameters temporal;
@@ -122,9 +123,9 @@ main(int argc, char ** argv)
     }
   }
 
-  ExaDG::GeneralParameters            general(input_file);
-  ExaDG::SpatialResolutionParameters  spatial(input_file);
-  ExaDG::TemporalResolutionParameters temporal(input_file);
+  ExaDG::GeneralParameters                 general(input_file);
+  ExaDG::SpatialResolutionParametersMinMax spatial(input_file);
+  ExaDG::TemporalResolutionParameters      temporal(input_file);
 
   // k-refinement
   for(unsigned int degree = spatial.degree_min; degree <= spatial.degree_max; ++degree)
