+++ include/exadg/convection_diffusion/driver.cpp
@@ -28,8 +28,8 @@
 #include <exadg/convection_diffusion/driver.h>
 #include <exadg/convection_diffusion/time_integration/create_time_integrator.h>
 #include <exadg/grid/get_dynamic_mapping.h>
+#include <exadg/operators/throughput_parameters.h>
 #include <exadg/utilities/print_solver_results.h>
-#include <exadg/utilities/throughput_parameters.h>
 
 namespace ExaDG
 {
