+++ include/exadg/convection_diffusion/solver.h
@@ -30,9 +30,10 @@
 #include <exadg/convection_diffusion/driver.h>
 
 // utilities
+#include <exadg/operators/resolution_parameters.h>
+#include <exadg/time_integration/resolution_parameters.h>
 #include <exadg/utilities/enum_patterns.h>
 #include <exadg/utilities/general_parameters.h>
-#include <exadg/utilities/resolution_parameters.h>
 
 // application
 #include <exadg/convection_diffusion/user_interface/declare_get_application.h>
@@ -47,7 +48,7 @@ create_input_file(std::string const & input_file)
   GeneralParameters general;
   general.add_parameters(prm);
 
-  SpatialResolutionParameters spatial;
+  SpatialResolutionParametersMinMax spatial;
   spatial.add_parameters(prm);
 
   TemporalResolutionParameters temporal;
@@ -128,9 +129,9 @@ main(int argc, char ** argv)
     }
   }
 
-  ExaDG::GeneralParameters            general(input_file);
-  ExaDG::SpatialResolutionParameters  spatial(input_file);
-  ExaDG::TemporalResolutionParameters temporal(input_file);
+  ExaDG::GeneralParameters                 general(input_file);
+  ExaDG::SpatialResolutionParametersMinMax spatial(input_file);
+  ExaDG::TemporalResolutionParameters      temporal(input_file);
 
   // k-refinement
   for(unsigned int degree = spatial.degree_min; degree <= spatial.degree_max; ++degree)
