+++ include/exadg/incompressible_navier_stokes/throughput.h
@@ -33,9 +33,9 @@
 #include <exadg/incompressible_navier_stokes/driver.h>
 
 // utilities
+#include <exadg/operators/hypercube_resolution_parameters.h>
+#include <exadg/operators/throughput_parameters.h>
 #include <exadg/utilities/general_parameters.h>
-#include <exadg/utilities/hypercube_resolution_parameters.h>
-#include <exadg/utilities/throughput_parameters.h>
 
 // application
 #include <exadg/incompressible_navier_stokes/user_interface/declare_get_application.h>
