+++ include/exadg/operators/hypercube_resolution_parameters.h
@@ -19,8 +19,8 @@
  *  ______________________________________________________________________
  */
 
-#ifndef INCLUDE_EXADG_UTILITIES_HYPERCUBE_RESOLUTION_PARAMETERS_H_
-#define INCLUDE_EXADG_UTILITIES_HYPERCUBE_RESOLUTION_PARAMETERS_H_
+#ifndef INCLUDE_EXADG_OPERATORS_HYPERCUBE_RESOLUTION_PARAMETERS_H_
+#define INCLUDE_EXADG_OPERATORS_HYPERCUBE_RESOLUTION_PARAMETERS_H_
 
 #include <deal.II/base/parameter_handler.h>
 
@@ -293,4 +293,4 @@ struct HypercubeResolutionParameters
 } // namespace ExaDG
 
 
-#endif /* INCLUDE_EXADG_UTILITIES_HYPERCUBE_RESOLUTION_PARAMETERS_H_ */
+#endif /* INCLUDE_EXADG_OPERATORS_HYPERCUBE_RESOLUTION_PARAMETERS_H_ */
