+++ include/exadg/poisson/overset_grids/driver.cpp
@@ -23,7 +23,6 @@
 #include <exadg/poisson/overset_grids/driver.h>
 #include <exadg/utilities/print_general_infos.h>
 #include <exadg/utilities/print_solver_results.h>
-#include <exadg/utilities/throughput_parameters.h>
 
 namespace ExaDG
 {
