+++ include/exadg/poisson/solver.h
@@ -28,8 +28,8 @@
 #include <exadg/poisson/driver.h>
 
 // utilities
+#include <exadg/operators/hypercube_resolution_parameters.h>
 #include <exadg/utilities/general_parameters.h>
-#include <exadg/utilities/hypercube_resolution_parameters.h>
 
 // application
 #include <exadg/poisson/user_interface/declare_get_application.h>
